requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
Flask-GraphQL==2.0.0
google-generativeai==0.8.4
graphene==2.1.9
gunicorn==21.2.0
python-dotenv==0.21.1
requests==2.28.1
beautifulsoup4==4.11.1
scikit-learn==1.1.2